Refactor dashboard into reusable modules
